--- src/nom_parser/mod.rs
@@ -1,3 +1,0 @@
-mod ast;
-mod parser;
-mod token;
