--- src/nom_parser/parser.rs
@@ -1,6 +1,0 @@
-pub mod parse_let;
-pub mod tools;
-
-fn parseExpression() {}
-
-fn parseStatemant() {}
