+++ src/parser/mod.rs
@@ -1,0 +1,5 @@
+mod ast;
+mod parse_let;
+mod tools;
+
+pub fn parse() {}
